requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
et_xmlfile==2.0.0
Flask==3.1.3
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
gurobipy==12.0.3
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==6.8.0
PyPDF2==3.0.1
pypdfium2==5.6.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tqdm==4.67.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.6
xlsxwriter==3.2.9
